Module collection::meta

Metadata for collections

Reexports

pub use meta::cardinality::Cardinality;

pub use meta::checksum::CheckSum;

pub use meta::max::Max;

pub use meta::key::Key;

Modules

cardinality
Metadata keeping track of the number of elements in the collection.

checksum
Metadata keeping track of the checksum of elements in the collection.

key
Metadata keeping track of the maximum key of elements in the collection.

max
Metadata keeping track of the maximum element in the collection.

Enums

Found
Like Selection but used on the branch level.

Selection
This enum is returned from select methods, to signal if something was found.

Traits

Meta
Metadata for T

Select
A trait that metadata can implement to allow searching through the tree.

SubMeta
Implemented for compound-Metadata, for each of the sub-metadatas.