Module collection::meta::key

Metadata keeping track of the maximum key of elements in the collection.

Implements Select over sorted collections, to find where a key would sort.

Structs

Key
A key, K is T::Key where T: Keyed

KeySum
The checksum of the hashes of all keys in collection

ValSum
The checksum of the hashes of all values in collection

Traits

Keyed
This T can be viewed as a Key-Value pair.