Struct collection::Collection

pub struct Collection<T, M, H, B> where H: CryptoHash {
    pub root: Location<H>,
    pub meta: Option<M>,
    pub freezer: Freezer<Node<T, M, H>, H, B>,
    pub divisor: usize,
}

A collection of elements of type T, with metadata of type M.

This is the base type, that all the collection operations are implemented over.

Fields
root: Location<H>

The location that constitutes the root for this collection.

meta: Option<M>

Top level metadata

freezer: Freezer<Node<T, M, H>, H, B>

The store for nodes

divisor: usize

The branching factor, currently hard-coded to 2, which means on average every fourth element will have weight > 0.

Methods

impl<T, M, H, B> Collection<T, M, H, B> where T: Weight + Freeze<H>,
M: Meta<T> + Freeze<H>,
H: CryptoHash,
B: Backend<Node<T, M, H>, H>

fn new(backend: B) -> Self

Returns a new, empty Collection.

fn new_from(root: Location<H>,
freezer: Freezer<Node<T, M, H>, H, B>)
-> Result<Self>

Constructs a Collection given a root, and a freezer

fn with_freezer(freezer: Freezer<Node<T, M, H>, H, B>) -> Self

Constructs a Collection given a freezer

fn meta(&self) -> &Option<M>

Returns the metadata of the collection, if non-empty

fn new_root(&mut self, loc: Location<H>) -> Result<()>

Sets the root of the collection, updating metadata

fn union_using<F, E>(&mut self, b: &mut Self) -> Result<Self> where F: Meta<T> + Select<T> + PartialEq + Ord,
E: Meta<T> + PartialEq,
M: SubMeta<F> + SubMeta<E>,
B: Backend<Node<T, M, H>, H> + Clone

Returns a new, cloned collection that is the result of a union operation given two Meta implementations F and E

F is used to select which T goes first in the union.

E is used to find overlapping common subtrees.

For Sets: F: Max<T>, E: CheckSum<T>, and for Maps: F: Key<T::Key>, E: KeySum<T>.

When the equality testing succeeds, elements will be picked from the Collection b.

fn persist(&self) -> Result<H::Digest>

Persists the collection to the underlying Backend

fn restore(hash: H::Digest, backend: B) -> Result<Self>

Restores a collection from the Backend given a hash

impl<T, M, H, B> Collection<T, M, H, B> where H: CryptoHash,
T: Weight + Freeze<H>,
M: Meta<T> + Freeze<H>,
B: Backend<Node<T, M, H>, H>

fn iter<'a>(&'a self) -> Iter<'a, T, M, Beginning, H, B>

Returns an iterator over Collection

fn iter_rev<'a>(&'a self) -> Iter<'a, T, M, End, H, B>

Returns a reverse iterator over Collection