Trait collection::MapOps

pub trait MapOps<K, V, M> where Self: Sized,
               M: Meta<KV<K, V>>,
               KV<K, V>: Keyed,
               KV<K, V>::Key: Weight + Clone + Ord,
               KV<K, V>::Value: Clone {
    fn insert(&mut self, key: KV<K, V>::Key, val: KV<K, V>::Value) -> Result<()>;
    fn remove(&mut self, key: KV<K, V>::Key) -> Result<Option<KV<K, V>::Value>>;
    fn get(&self, key: KV<K, V>::Key) -> Result<Option<Cow<KV<K, V>::Value>>>;
    fn mutate<F>(&mut self, key: KV<K, V>::Key, f: F) -> Result<Option<()>> where F: FnOnce(&mut KV<K, V>::Value);
}

Map operations on a Collection

Required Methods

fn insert(&mut self, key: KV<K, V>::Key, val: KV<K, V>::Value) -> Result<()>

Insert a value V at key K

fn remove(&mut self, key: KV<K, V>::Key) -> Result<Option<KV<K, V>::Value>>

Remove value at key K

fn get(&self, key: KV<K, V>::Key) -> Result<Option<Cow<KV<K, V>::Value>>>

Get a reference to the value at key K

fn mutate<F>(&mut self, key: KV<K, V>::Key, f: F) -> Result<Option<()>> where F: FnOnce(&mut KV<K, V>::Value)

Mutate the value at key K with function F