Crate collection

Persistent datastructure

Reexports

pub use meta::Meta;

pub use meta::SubMeta;

pub use meta::Max;

pub use meta::CheckSum;

pub use meta::Key;

pub use meta::Cardinality;

Modules

meta
Metadata for collections

Macros

collection

Structs

BlakeWrap
Wraps Blake2b in the CryptoHash trait

Collection
A collection of elements of type T, with metadata of type M.

Source
A source to read bytes from

Traits

CryptoHash
A wrapper trait for a cryptographic hash function

Freeze
Write &self as bytes or read Self from bytes

MapOps
Map operations on a Collection

MapOpsKeySum
Operations on a map with KeySum metadata

SetOps
Set operations on a Collection

SetOpsCheckSum
Set operations on Checksummed sets

Sink
A sink implementation

VectorOps
Vector Operations on a Collection

Weight
The "weight" of an element, should produce a uniform distribution from the possible values of the type implementing this trait.